fear and loathing in API HELL
