+++ Pipfile
@@ -17,6 +17,7 @@ pytest-mock = "*"
 pytest-splinter = "*"
 flask-pymongo = "*"
 bcrypt = "*"
+flask-basicauth = "*"
 
 [dev-packages]
 pytest = "*"
